// src/modules/transfer/types/transferring-status.type.ts
export enum TRANSFERRING_STATUS {
	AVAILABLE = 'AVAILABLE',
	WAIT_TRANSFER_ACCEPTED = 'WAIT_TRANSFER_ACCEPTED',
	TRANSFER_SUCCESS = 'TRANSFER_SUCCESS',
	TRANSFER_FAILED = 'TRANSFER_FAILED',
	CANCEL_TRANSFERRING = 'CANCEL_TRANSFERRING',
	TRANSFERING = 'TRANSFERING',
	WAIT_FOR_RECIPIENT_RECEIVE_FILE = 'WAIT_FOR_RECIPIENT_RECEIVE_FILE',
	REFUSE_REQUEST = 'REFUSE_REQUEST',
}